requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
discord.py==2.1.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
Flask-HTTPAuth==4.7.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gunicorn==20.1.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
ipywidgets==8.0.2
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.4
jupyter_core==5.0.0
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.3
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.3
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
pycryptodomex==3.15.0
Pygments==2.13.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
stack-data==0.6.0
termcolor==2.1.0
tornado==6.2
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.3
yarg==0.1.9
yarl==1.8.2
